code/llm_api_middleware/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pyyaml>=6.0
